Added intermediate processes for supervision
It seems like the supervisors expect otp modules for supervision and
not normal processes which is why I probably should have used a gen_fsm
for client handling. Instead we have an intermediate process which
spawns each acceptor and links to it. If it crashes the supervisor
restarts the intermediate which restarts the acceptor.
This means that the acceptor_supervisor should initialize properly this
time.